max/min value
Resolve #1
