--- test_max_min.py
@@ -1,13 +1,0 @@
-from demo import *
-import random
-
-
-def test_max_min():
-    arr = []
-    for n in range(10):
-        arr.append(random.randint(10, 100))
-
-    m = max_min(arr)
-
-    for x in arr:
-        assert(m >= x)
